refactor: move expo manifest to parsers for new native parsers
